Usando Endponit para ordenar as Tarefas.

+++ lista_de_tarefas/src/main/java/com/fatto/lista_de_tarefas/model/TarefaDTO.java
@@ -2,4 +2,4 @@ package com.fatto.lista_de_tarefas.model;
 
 import java.time.LocalDate;
 
-public record TarefaDTO(String nometarefa, float custo, LocalDate datalimite, Integer ordem) {}
+public record TarefaDTO(String nometarefa, float custo, LocalDate datalimite, Long ordem) {}
